really skip tests when config not present

+++ tests/unit/PostgresAdapterTest.php
@@ -36,6 +36,9 @@ class PostgresAdapterTest extends PHPUnit_Framework_TestCase
      */
     protected function tearDown()
     {
+        if(!$this->adapter)
+          return;
+        
         //delete any tables we created
         if ($this->adapter->has_table('users',true)) {
             $this->adapter->drop_table('users');
